requirements.txt
aiogram==3.21.0
python-dotenv